# internal/httpserver/swaggerdocs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "GoDNS API",
        "contact": {
            "name": "GoDNS Support",
            "url": "https://github.com/rogerwesterbo/godns",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/rogerwesterbo/godns/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:14000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/zones": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get a list of all DNS zones",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "List all DNS zones",
                "responses": {
                    "200": {
                        "description": "List of zones",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSZone"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Create a new DNS zone with optional records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Create a new DNS zone",
                "parameters": [
                    {
                        "description": "Zone to create",
                        "name": "zone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSZone"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Zone created",
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSZone"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Zone already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/zones/{zone}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get a specific DNS zone by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Get a DNS zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone details",
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSZone"
                        }
                    },
                    "404": {
                        "description": "Zone not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Update an existing DNS zone (replaces all records)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Update a DNS zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated zone data",
                        "name": "zone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSZone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSZone"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Zone not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Delete a DNS zone and all its records",
                "tags": [
                    "Zones"
                ],
                "summary": "Delete a DNS zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zone deleted"
                    },
                    "404": {
                        "description": "Zone not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_rogerwesterbo_godns_internal_models.DNSRecord": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Fully qualified domain name",
                    "type": "string",
                    "example": "www.example.lan."
                },
                "ttl": {
                    "description": "Time to live in seconds",
                    "type": "integer",
                    "example": 300
                },
                "type": {
                    "description": "A, AAAA, MX, NS, TXT, etc.",
                    "type": "string",
                    "example": "A"
                },
                "value": {
                    "description": "The record value (IP, hostname, etc.)",
                    "type": "string",
                    "example": "192.168.1.100"
                }
            }
        },
        "github_com_rogerwesterbo_godns_internal_models.DNSZone": {
            "type": "object",
            "properties": {
                "domain": {
                    "description": "e.g., \"example.lan.\"",
                    "type": "string",
                    "example": "example.lan."
                },
                "records": {
                    "description": "DNS records in this zone",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rogerwesterbo_godns_internal_models.DNSRecord"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:14101/realms/godns/protocol/openid-connect/token"
        }
    },
...
}